chore: renaming config and readme updates
