+++ src/app/stages/application.stage.ts
@@ -2,7 +2,7 @@ import * as cdk from "aws-cdk-lib";
 import { Construct } from "constructs";
 import { StatefulStack } from "../stacks/stateful.stack";
 import { StatelessStack } from "../stacks/stateless.stack";
-import { PROJECT_NAME } from "../../config/config";
+import { PROJECT_NAME } from "../../config";
 import { EnvironmentConfig } from "../../config/types";
 
 export class ApplicationStage extends cdk.Stage {
