+++ src/feature-dev.ts
@@ -3,7 +3,7 @@ import * as cdk from "aws-cdk-lib";
 import * as dotenv from "dotenv";
 import { StatefulStack } from "./app/stacks/stateful.stack";
 import { StatelessStack } from "./app/stacks/stateless.stack";
-import { environments, PROJECT_NAME } from "./config/config";
+import { environments, PROJECT_NAME } from "./config";
 import { Stage } from "./config/types";
 
 dotenv.config();
