+++ src/main.ts
@@ -1,5 +1,5 @@
 import * as cdk from "aws-cdk-lib";
-import { pipelineConfig, PROJECT_NAME } from "./config/config";
+import { pipelineConfig, PROJECT_NAME } from "./config";
 import { PipelineStack } from "./pipeline/pipeline.stack";
 
 const app = new cdk.App();
